# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ConsolidationRequest": {
        "additionalProperties": false,
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/ExecutionMode",
            "default": "demo"
          },
          "scope_id": {
            "format": "uuid",
            "title": "Scope Id",
            "type": "string"
          },
          "source_memory_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "maxItems": 8,
            "minItems": 3,
            "title": "Source Memory Ids",
            "type": "array"
          }
        },
        "required": [
          "scope_id",
          "source_memory_ids"
        ],
        "title": "ConsolidationRequest",
        "type": "object"
      },
...
      "ExecutionMode": {
        "enum": [
          "demo",
          "live"
        ],
        "title": "ExecutionMode",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IngestInteractionRequest": {
        "additionalProperties": false,
        "properties": {
          "idempotency_key": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          },
          "mode": {
            "$ref": "#/components/schemas/ExecutionMode",
            "default": "demo"
          },
          "occurred_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Occurred At"
          },
          "preview": {
            "default": false,
            "title": "Preview",
            "type": "boolean"
          },
          "scope_id": {
            "format": "uuid",
            "title": "Scope Id",
            "type": "string"
          },
          "source_ref": {
            "anyOf": [
              {
                "maxLength": 512,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Ref"
          },
          "text": {
            "maxLength": 20000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "scope_id",
          "text"
        ],
        "title": "IngestInteractionRequest",
        "type": "object"
      },
...
      "MemoryType": {
        "enum": [
          "preference",
          "semantic",
          "episodic",
          "procedural"
        ],
        "title": "MemoryType",
        "type": "string"
      },
...
      "ReviewItem": {
        "additionalProperties": false,
        "properties": {
          "candidate": {
            "$ref": "#/components/schemas/CandidateMemory"
          },
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "evidence_excerpt": {
            "title": "Evidence Excerpt",
            "type": "string"
          },
          "existing_memory": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MemoryRecord"
              },
              {
                "type": "null"
              }
            ]
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "conflict",
              "consolidation"
            ],
            "title": "Kind",
            "type": "string"
          },
          "memory_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Memory Id"
          },
          "proposed_relation": {
            "$ref": "#/components/schemas/MemoryRelation"
          },
          "reason_code": {
            "title": "Reason Code",
            "type": "string"
          },
          "reason_summary": {
            "title": "Reason Summary",
            "type": "string"
          },
          "resolution": {
            "anyOf": [
              {
                "enum": [
                  "keep_both",
                  "use_new",
                  "keep_existing",
                  "invalid"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution"
          },
          "resolution_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Reason"
          },
          "resolved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "scope_id": {
            "format": "uuid",
            "title": "Scope Id",
            "type": "string"
          },
          "source_memories": {
            "items": {
              "$ref": "#/components/schemas/MemoryRecord"
            },
            "title": "Source Memories",
            "type": "array"
          },
          "status": {
            "enum": [
              "pending",
              "resolved"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "scope_id",
          "kind",
          "status",
          "candidate",
          "existing_memory",
          "source_memories",
          "proposed_relation",
          "confidence",
          "evidence_excerpt",
          "reason_code",
          "reason_summary",
          "created_at",
          "resolved_at",
          "resolution",
          "resolution_reason",
          "memory_id"
        ],
        "title": "ReviewItem",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Explainable long-term memory infrastructure for AI agents.",
    "title": "MemoryOS API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/consolidations/propose": {
      "post": {
        "operationId": "propose_consolidation_v1_consolidations_propose_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsolidationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Propose Consolidation",
        "tags": [
          "v1"
        ]
      }
    },
...
    "/v1/interactions": {
      "post": {
        "operationId": "ingest_interaction_v1_interactions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestInteractionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestInteractionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ingest Interaction",
        "tags": [
          "v1"
        ]
      }
    },
    "/v1/interactions/{interaction_id}": {
      "get": {
        "operationId": "get_interaction_v1_interactions__interaction_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "interaction_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Interaction Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scope_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scope Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestInteractionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Interaction",
        "tags": [
          "v1"
        ]
      }
    },
    "/v1/memories": {
      "get": {
        "operationId": "list_memories_v1_memories_get",
        "parameters": [
          {
            "in": "query",
            "name": "scope_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scope Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "memory_types",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/MemoryType"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Memory Types"
            }
          },
          {
            "in": "query",
            "name": "statuses",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/MemoryStatus"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Statuses"
            }
          },
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Memories",
        "tags": [
          "v1"
        ]
      }
    },
    "/v1/memories/{memory_id}": {
      "get": {
        "operationId": "get_memory_v1_memories__memory_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "memory_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Memory Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scope_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scope Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Memory",
        "tags": [
          "v1"
        ]
      }
    },
    "/v1/memories/{memory_id}/forget": {
      "post": {
        "operationId": "forget_memory_v1_memories__memory_id__forget_post",
        "parameters": [
          {
            "in": "path",
            "name": "memory_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Memory Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scope_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scope Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgetMemoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForgetMemoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Forget Memory",
        "tags": [
          "v1"
        ]
      }
    },
    "/v1/memories/{memory_id}/history": {
      "get": {
        "operationId": "memory_history_v1_memories__memory_id__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "memory_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Memory Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scope_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scope Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Memory History",
        "tags": [
          "v1"
        ]
      }
    },
    "/v1/memories/{memory_id}/resolve": {
      "post": {
        "operationId": "resolve_memory_v1_memories__memory_id__resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "memory_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Memory Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scope_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scope Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveMemoryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveMemoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Memory",
        "tags": [
          "v1"
        ]
      }
    },
    "/v1/overview": {
      "get": {
        "operationId": "overview_v1_overview_get",
        "parameters": [
          {
            "in": "query",
            "name": "scope_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scope Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OverviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Overview",
        "tags": [
          "v1"
        ]
      }
    },
    "/v1/recall": {
      "post": {
        "operationId": "recall_v1_recall_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecallRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecallResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recall",
        "tags": [
          "v1"
        ]
      }
    },
    "/v1/recall/compare": {
      "post": {
        "operationId": "compare_recall_v1_recall_compare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecallRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecallComparisonResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compare Recall",
        "tags": [
          "v1"
        ]
      }
    },
    "/v1/reviews": {
      "get": {
        "operationId": "list_reviews_v1_reviews_get",
        "parameters": [
          {
            "in": "query",
            "name": "scope_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scope Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "pending",
              "enum": [
                "pending",
                "resolved"
              ],
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Reviews",
        "tags": [
          "v1"
        ]
      }
    },
    "/v1/reviews/{review_id}/resolve": {
      "post": {
        "operationId": "resolve_review_v1_reviews__review_id__resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "review_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Review Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "scope_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Scope Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveReviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewItem"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Review",
        "tags": [
          "v1"
        ]
      }
    }
  }
}